cleanups
A bunch of typos.
Xterm on Linux and iTerm 2 on macOS are good enough (and often
the most/only supported terminals).
Spectacle replaces chunkwm (now https://github.com/koekeishiya/yabai)
and skhd. At the time chunkwm wasn't as good as bspwm (and sometimes
slighly annoying) and I'm not sure what its current state is (disabling
System Integrity Protection is kind of a no go).

Also update everything because it's been a while and because I
encountered
https://lists.gnu.org/archive/html/emacs-devel/2020-05/msg02859.html